+++ tests/test_vector.c
@@ -21,8 +21,8 @@ static void test_fill_vector(void)
     gv_insert_elem_at_pos(&vector, 11, 0);
     gv_insert_elem_at_pos(&vector, 12, 7);
 
-    ASSERT_EQUAL(vector.len, 8);
-    ASSERT_EQUAL(vector.size, 8);
+    ASSERT_EQUAL((int)vector.len, 8);
+    ASSERT_EQUAL((int)vector.size, 8);
 
     gv_for_each_pos(pos, &vector) {
         ASSERT_EQUAL(vector.tab[pos], tab[idx_tab]);
@@ -53,8 +53,8 @@ static void test_vector_add_and_remove_elem(void)
     gv_insert_elem_at_pos(&vector, 12, 4);
     gv_remove(&vector, 3);
 
-    ASSERT_EQUAL(vector.len, 4);
-    ASSERT_EQUAL(vector.size, 6);
+    ASSERT_EQUAL((int)vector.len, 4);
+    ASSERT_EQUAL((int)vector.size, 6);
 
     gv_for_each_pos(pos, &vector) {
         ASSERT_EQUAL(vector.tab[pos], tab[idx_tab]);
@@ -86,8 +86,8 @@ static void test_fill_vector_sorting(void)
     gv_insert_elem_sorted(&vector, 4, cmp_elem);
     gv_insert_elem_sorted(&vector, 2, cmp_elem);
 
-    ASSERT_EQUAL(vector.len, 5);
-    ASSERT_EQUAL(vector.size, 5);
+    ASSERT_EQUAL((int)vector.len, 5);
+    ASSERT_EQUAL((int)vector.size, 5);
 
     gv_for_each_pos(pos, &vector) {
         ASSERT_EQUAL(vector.tab[pos], tab[idx_tab]);
@@ -113,8 +113,8 @@ static void test_sort_vector_simple(void)
 
     gv_sort(&vector, cmp_elem);
 
-    ASSERT_EQUAL(vector.len, 5);
-    ASSERT_EQUAL(vector.size, 5);
+    ASSERT_EQUAL((int)vector.len, 5);
+    ASSERT_EQUAL((int)vector.size, 5);
 
     gv_for_each_pos(pos, &vector) {
         ASSERT_EQUAL(vector.tab[pos], tab[idx_tab]);
@@ -219,8 +219,8 @@ static void test_fill_pointer_vector(void)
     add_new_elem_and_set(&vector, 4);
     add_new_elem_and_set(&vector, 5);
 
-    ASSERT_EQUAL(vector.len, 5);
-    ASSERT_EQUAL(vector.size, 5);
+    ASSERT_EQUAL((int)vector.len, 5);
+    ASSERT_EQUAL((int)vector.size, 5);
 
     gv_for_each_pos(pos, &vector) {
         ASSERT_EQUAL(*(vector.tab[pos]), tab[idx_tab]);
@@ -243,8 +243,8 @@ static void test_new_and_delete_vector(void)
     gv_add(vector, 4);
     gv_add(vector, 5);
 
-    ASSERT_EQUAL(vector->len, 5);
-    ASSERT_EQUAL(vector->size, 5);
+    ASSERT_EQUAL((int)vector->len, 5);
+    ASSERT_EQUAL((int)vector->size, 5);
 
     gv_for_each_pos(pos, vector) {
         ASSERT_EQUAL(vector->tab[pos], tab[idx_tab]);
@@ -270,8 +270,8 @@ static void test_vector_init_size(void)
     gv_insert_elem_at_pos(&vector, 11, 0);
     gv_insert_elem_at_pos(&vector, 12, 7);
 
-    ASSERT_EQUAL(vector.len, 8);
-    ASSERT_EQUAL(vector.size, 10);
+    ASSERT_EQUAL((int)vector.len, 8);
+    ASSERT_EQUAL((int)vector.size, 10);
 
     gv_for_each_pos(pos, &vector) {
         ASSERT_EQUAL(vector.tab[pos], tab[idx_tab]);
@@ -328,8 +328,8 @@ static void test_reset_vector(void)
     gv_add(&vector, 7);
     gv_add(&vector, 8);
 
-    ASSERT_EQUAL(vector.len, 3);
-    ASSERT_EQUAL(vector.size, 5);
+    ASSERT_EQUAL((int)vector.len, 3);
+    ASSERT_EQUAL((int)vector.size, 5);
 
     gv_for_each_pos(pos, &vector) {
         ASSERT_EQUAL(vector.tab[pos], tab[idx_tab]);
